FSM & States: I've made ALLOT of improvements to the State Machine.  It now supports infinite changing and regression of those changes upon completion, if desired. I added in some more standard states such as Dead and Attack Mob.  Dead of course does not work all the way as LUA is still not working, but it will use PPather to compute a path to your corpse and then call the MoveTo State to get you there.  I also created the start of a "Base Global State" called GlobalBaseBotState which will attempt to be the base for all global states.
Scripting:
I have updated the scripting files so that it does not use the interfaces that it was using before and now a script will be a State instead.
I did some testing and I was able to inherit from State and even use other states to perform actions.  So scripts will now be going to states (global ones anyways).
Or at least thats where I think it will head.  That way users can write their own global states.

Clean-up:
I cleaned up allot of functions such as the math functions, which were mixed around here and there (includeing some in the wowplayer file).
I have tried to put all of these in the MathFuncs class.  I also added in the missing location information for WoWDynamicObject.
The timeout for generating a path on PPather has been increased from 15 seconds to 120 seconds, just for testing purposes as generating
a path from the dwarf starting area to Ironforge was taking a bit of time :).

Facing Calculations and Code:
The code that would face you towards a destination was touchy at best so I have added in some code that uses memory writing to help improve
it.  Rright now it has two modes: #1 is a non timed rotation, so it just writes the new orientation to the memory and all at once the character is
now moving that direction.  This works great but it totally looks like a bot of course.  I was able to travel all the way from the starting area
for dwarfs to the back of the bank in Ironforge following saved waypoints using this method.  The other way, using the same function but with
the timed parameter set to true instead of false, will attempt to rotate your character in the right direction.  It works most of the time, but may
still need some work.